Use the proper type for slotsize.

+++ src/swap.cpp
@@ -39,7 +39,7 @@ static char *cbuf;			/* sector buffer */
 static sector cached;			/* sector currently cached in cbuf */
 static header *first, *last;		/* first and last swap slot */
 static header *lfree;			/* free swap slot list */
-static size_t slotsize;			/* sizeof(header) + size of sector */
+static off_t slotsize;			/* sizeof(header) + size of sector */
 static unsigned int sectorsize;		/* size of sector */
 static unsigned int restoresecsize;	/* size of sector in restore file */
 static sector swapsize, cachesize;	/* # of sectors in swap and cache */
